+++ generators/PlnX86_64Generator.h
@@ -28,7 +28,7 @@ public:
 	void genMul(PlnGenEntity* tgt, PlnGenEntity* second);
 	void genDiv(PlnGenEntity* tgt, PlnGenEntity* second, string comment);
 
-	unique_ptr<PlnGenEntity> getInt(int i);
+	unique_ptr<PlnGenEntity> getInt(int64_t i);
 	unique_ptr<PlnGenEntity> getStackAddress(int offset, int size);
 	unique_ptr<PlnGenEntity> getStrAddress(int index);
 	unique_ptr<PlnGenEntity> getArgument(int i, int size);
